configure: Improve readline detection
Detecting readline properly is surprisingly challenging, because
of many factors: for starters, the library only started shipping
a pkg-config file somewhat recently, and even when that happened
the file turned out to be broken on non-Linux platforms; in
addition to that, macOS complicates things further by including
a library which looks like it's intended to be a drop-in
replacement for readline, since it declares the same symbols, but
in reality doesn't support many of the modern features and is
thus unusable for our purposes.

The detection code is lifted almost verbatim from libvirt.

Signed-off-by: Andrea Bolognani <eof@kiyuko.org>