docs: Rewrite man page in reStructuredText
pandoc, which we've been using to generate the man page up until
now, works well but is a pretty big hammer for this specific
nail; additionally, due to the fact that it's implemented in
Haskell, it's often not available on all architectures.

rst2man, part of Python Docutils, is a much smaller tool that's
written in Python and as such runs pretty much everywhere.

Signed-off-by: Andrea Bolognani <eof@kiyuko.org>